Refactored re-index to delete lucene index files first, and then just add which should be at least 2.5x  faster than the old way.  Also now index ControlledVocabulary terms

+++ app/ix/core/adapters/EntityPersistAdapter.java
@@ -23,7 +23,6 @@ import javax.persistence.PreUpdate;
 import org.apache.lucene.store.AlreadyClosedException;
 
 import com.avaje.ebean.event.BeanPersistAdapter;
-import com.avaje.ebean.event.BeanPersistListener;
 import com.avaje.ebean.event.BeanPersistRequest;
 
 import ix.core.EntityProcessor;
@@ -40,7 +39,6 @@ import ix.core.plugins.SequenceIndexerPlugin;
 import ix.core.plugins.StructureIndexerPlugin;
 import ix.core.plugins.TextIndexerPlugin;
 import ix.core.processors.BackupProcessor;
-import ix.ginas.models.v1.Substance;
 import ix.seqaln.SequenceIndexer;
 import ix.utils.EntityUtils;
 import ix.utils.TimeProfiler;
@@ -75,7 +73,7 @@ public class EntityPersistAdapter extends BeanPersistAdapter{
     
     
     
-    private TextIndexerPlugin plugin = 
+    private TextIndexerPlugin textIndexerPlugin =
             Play.application().plugin(TextIndexerPlugin.class);
     private static StructureIndexerPlugin strucProcessPlugin;
     private static SequenceIndexerPlugin seqProcessPlugin;
@@ -402,8 +400,8 @@ public class EntityPersistAdapter extends BeanPersistAdapter{
     }
     
 	private void makeIndexOnBean(Object bean) throws java.io.IOException {
-		if (plugin != null){
-			plugin.getIndexer().add(bean);
+		if (textIndexerPlugin != null){
+			textIndexerPlugin.getIndexer().add(bean);
 		}
 
 		List<Field> sequenceFields = getSequenceIndexableField(bean);
@@ -438,8 +436,8 @@ public class EntityPersistAdapter extends BeanPersistAdapter{
 	}
 	
 	private void deleteIndexOnBean(Object bean) throws Exception {
-		if (plugin != null)
-            plugin.getIndexer().remove(bean);
+		if (textIndexerPlugin != null)
+            textIndexerPlugin.getIndexer().remove(bean);
 		String _id = EntityUtils.getIdForBeanAsString(bean);
 		List<Field> sequenceFields = getSequenceIndexableField(bean);
 		if (sequenceFields != null && sequenceFields.size()>0) {
@@ -642,8 +640,10 @@ public class EntityPersistAdapter extends BeanPersistAdapter{
             }
         }
     }
-    
     public void deepreindex(Object bean){
+        deepreindex(bean, true);
+    }
+    public void deepreindex(Object bean, boolean deleteFirst){
     	//reindex(bean);
     	
     	if(bean instanceof Model){
@@ -651,7 +651,7 @@ public class EntityPersistAdapter extends BeanPersistAdapter{
 				@Override
 				public void call(Object m, String path) {
 						
-					reindex(m);
+					reindex(m, deleteFirst);
 				}
 	    	});
     	}
@@ -660,7 +660,10 @@ public class EntityPersistAdapter extends BeanPersistAdapter{
     
 //    public static long reindexCount=0;
 //    public Stack<Long> times= new Stack<Long>();
-    public void reindex(Object bean){
+public void reindex(Object bean){
+    reindex(bean, true);
+}
+    public void reindex(Object bean, boolean deleteFirst){
     	
     	String _id=null;
     	if(bean instanceof BaseModel){
@@ -675,9 +678,12 @@ public class EntityPersistAdapter extends BeanPersistAdapter{
         try {
 //        	long start=System.currentTimeMillis();
         	//times.push();
-            if(_id!=null)
-                alreadyLoaded.put(bean.getClass()+_id,_id);
-            deleteIndexOnBean(bean);
+            if(_id!=null) {
+                alreadyLoaded.put(bean.getClass() + _id, _id);
+            }
+            if(deleteFirst) {
+                deleteIndexOnBean(bean);
+            }
             makeIndexOnBean(bean);
             
 //            times.pop();
